+++ internal/storage/qdrant/semantic_cache_test.go
@@ -341,14 +341,28 @@ func filterHasField(filter *qdrant.Filter, key string) bool {
 }
 
 type recordingPointsClient struct {
-	deleteReq *qdrant.DeletePoints
+	upsertReq     *qdrant.UpsertPoints
+	setPayloadReq *qdrant.SetPayloadPoints
+	deleteReq     *qdrant.DeletePoints
+	searchReqs    []*qdrant.SearchPoints
+	searchFn      func(*qdrant.SearchPoints) (*qdrant.SearchResponse, error)
 }
 
-func (c *recordingPointsClient) Upsert(context.Context, *qdrant.UpsertPoints, ...grpc.CallOption) (*qdrant.PointsOperationResponse, error) {
+func (c *recordingPointsClient) Upsert(_ context.Context, req *qdrant.UpsertPoints, _ ...grpc.CallOption) (*qdrant.PointsOperationResponse, error) {
+	c.upsertReq = req
 	return &qdrant.PointsOperationResponse{}, nil
 }
 
-func (c *recordingPointsClient) Search(context.Context, *qdrant.SearchPoints, ...grpc.CallOption) (*qdrant.SearchResponse, error) {
+func (c *recordingPointsClient) SetPayload(_ context.Context, req *qdrant.SetPayloadPoints, _ ...grpc.CallOption) (*qdrant.PointsOperationResponse, error) {
+	c.setPayloadReq = req
+	return &qdrant.PointsOperationResponse{}, nil
+}
+
+func (c *recordingPointsClient) Search(_ context.Context, req *qdrant.SearchPoints, _ ...grpc.CallOption) (*qdrant.SearchResponse, error) {
+	c.searchReqs = append(c.searchReqs, req)
+	if c.searchFn != nil {
+		return c.searchFn(req)
+	}
 	return &qdrant.SearchResponse{}, nil
 }